ml-bug-prediction/requirements/requirements_webapp.txt
Flask==3.0.0
pandas>=2.2.0
jira==3.8.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dateutil>=2.8.2
numpy>=1.26.0
Werkzeug>=3.0.0
Jinja2>=3.1.2
requests>=2.31.0
python-dotenv>=1.0.0 